refactor: simplify xml serializer design
